+++ app/models/concerns/attr_accssor.rb
@@ -1,0 +1,7 @@
+module Attr_accssor
+  extend ActiveSupport::Concern
+
+  included do
+    attr_accessor :selectbox_parameter
+  end
+end
