+++ app/models/concerns/enum_status.rb
@@ -1,0 +1,7 @@
+module Enum_status
+  extend ActiveSupport::Concern
+
+  included do
+    enum status: { run: 0, done: 1, expired: 2 }
+  end
+end
